#28 - Seperate namespaceService from elementService
- Add NamespaceService
- Transfer all namespace related methods from elementService to namespaceService
